set the preset eq command for the ble device

+++ JBLHeadphone2/app/src/main/java/jbl/stc/com/entity/IIR_PARAM_T.java
@@ -1,7 +1,7 @@
 package jbl.stc.com.entity;
 
 public class IIR_PARAM_T{
-    public IIR_PARAM_T(int type,float gain, int fc, float Q){
+    public IIR_PARAM_T(int type,float gain, float fc, float Q){
         mType = type;
         mGain = gain;
         mFc = fc;
@@ -9,6 +9,6 @@ public class IIR_PARAM_T{
     }
     public int         mType;
     public float       mGain;
-    public int         mFc;
+    public float       mFc;
     public float       mQ;
 }
